requirements.txt
fastapi~=0.116.0
loguru~=0.7.3
uvicorn~=0.35.0
python-dotenv~=1.1.1
pydantic~=2.11.7
volcengine~=1.0.192